minor fixes.

+++ database/migrations/2022_09_15_180101_create_properties_table.php
@@ -19,9 +19,9 @@ return new class extends Migration
             $table->text("address");
             $table->text("image_urls");
             $table->decimal("value_usd", 11, 2);
-            $table->decimal("percentage_available", 11, 2)->default(75);
+            $table->decimal("percentage_available", 3, 0)->default(75);
             $table->decimal("dividend_usd", 11, 2)->default(0);
-            $table->decimal("size_sf", 11, 2);
+            $table->integer("size_sf");
             $table->timestampTz("created_at");
             $table->timestampTz("updated_at");
         });
